requirements.txt
docopt
sh
color-print==0.0.2